moved ThreadSafeLightData and related methods to StationaryLightsManager
